Adding tests to back up the UniquessValidator.
The UniquessValidator needs a full object. So it tries
to assign every field it knows about if the record responds to it.

Not a very pretty solution, but the tests pass for now. This piece of
code would probably need to change in the future as I'm not sure how
nested objects would respond to this.