Add Section to README on Virtus & Arrays of IDs
I think this is a not uncommon situation, and when I ran into it it took a little bit of research to find a good solution.  (I wanted to avoid the custom `schools=` method, since that seemed like a "last resort" solution.)

I think this technique has a lot of benefits, especially readability and the ability to easily add min/max testing.